pyproject.toml
[tool.poetry]
name = "passwordmanager"
version = "0.1.0"
description = ""
authors = ["edoardomanieri"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.3.1"
cffi = "1.14.4"
cryptography = "3.3.1"
django = "3.1.5"
django-cors-headers = "3.7.0"
djangorestframework = "3.12.2"
djangorestframework-jwt = "1.11.0"
pycparser = "2.20"
pyjwt = "1.7.1"
pymysql = "1.0.2"
pytz = "2020.5"
six = "1.15.0"
sqlparse = "0.4.1"
gunicorn = "20.1.0"
whitenoise = "5.3.0"
dj-database-url = "0.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
